Fix static analysis
